Move get_checked_document_id_for_view_id to validation
